+++ test/common/src/main/java/org/hibernate/tool/hbmlint/HbmLintTest/TestCase.java
@@ -3,10 +3,10 @@ package org.hibernate.tool.hbmlint.HbmLintTest;
 import java.io.File;
 
 import org.hibernate.tool.api.metadata.MetadataDescriptor;
-import org.hibernate.tool.hbm2x.HbmLintExporter;
 import org.hibernate.tool.internal.export.lint.BadCachingDetector;
 import org.hibernate.tool.internal.export.lint.Detector;
 import org.hibernate.tool.internal.export.lint.HbmLint;
+import org.hibernate.tool.internal.export.lint.HbmLintExporter;
 import org.hibernate.tool.internal.export.lint.InstrumentationDetector;
 import org.hibernate.tool.internal.export.lint.ShadowedIdentifierDetector;
 import org.hibernate.tools.test.util.HibernateUtil;
